+++ game/match/matchsystem.go
@@ -72,6 +72,11 @@ func (s *MatchSystem) NewMatch(matchName string, hostID string, visible bool) *M
 		spectators:        internal.NewConcurrentDictionary[Spectator](),
 		persistentEffects: make(map[int]PersistentEffect),
 		Turn:              1,
+		
+		PlayerSelectingToss: "",
+		TossOutcome: 	   0,
+		TossPrediction:    0,
+
 		Started:           false,
 		Visible:           visible,